requirements.txt
omegaconf
scikit-learn
numpy
timm
nltk
lmdb
Pillow
imgaug
hydra-core
fvcore
ray[tune]
ax-platform
PyYAML
tqdm
setuptools
opencv-python
opendatasets
transformers
seaborn
ftfy
pytorch-lightning==1.9
